requirements.txt
telebot~=0.0.5
pyTelegramBotAPI~=4.23.0
python-dotenv~=1.0.1